+++ commands/music/rewind.js
@@ -1,0 +1,5 @@
+module.exports = {
+  name: "rewind",
+  category: "music",
+  run: async (message, args, client, Discord, player, queue) => {},
+};