+++ commands/music/seek.js
@@ -1,0 +1,5 @@
+module.exports = {
+  name: "seek",
+  category: "music",
+  run: async (message, args, client, Discord, player, queue) => {},
+};