+++ changelog.md
@@ -5,7 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0.4] - 2021-05-05 [LATEST]
+## [0.5.0] - 2021-05-06 [LATEST]
+
+### Fixed
+
+- Versioning System
+- Typo in Filter.js
+
+### Added
+
+- Bot Activity (Status)
+- Fininshed Music Commands (queue,skipto)
+- Un Fininshed Music Commands (rewind,seek)
+- Bot Server Deafen For Smooth Playback
+- Dev Only Commnads (Not Working)
+
+### Updated
+
+- Music Comands and made more good
+- Command Handler for Dev Only Commands
+
+## [0.4.0] - 2021-05-05
 
 ### Added
 
@@ -15,7 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Version in Pacage.json
 
-## [0.0.3] - 2021-05-04
+## [0.3.0] - 2021-05-04
 
 ### Added
 
@@ -34,7 +54,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - client.prefix bug in ready event of bot
 
-## [0.0.2] - 2021-05-03
+## [0.2.0] - 2021-05-03
 
 ### Added
 
@@ -56,7 +76,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Refactored Skip, Stop, Play Command
 - Command Handler by passing Queue as an Param in Run() function.
 
-## [0.0.1] - 2021-04-27
+## [0.1.0] - 2021-04-27
 
 ### Added